Change:過度な共通化を排除、各々のWidgetテストファイルでbuildWidgetを持つように変更
