+++ test/search_bar_test.dart
@@ -3,14 +3,30 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:yumemi_codecheck_repo_search/generated/l10n.dart';
 import 'package:yumemi_codecheck_repo_search/page/widget/search_bar.dart';
-import 'package:yumemi_codecheck_repo_search/provider/search_query_provider.dart';
 import 'package:yumemi_codecheck_repo_search/provider/service_provider.dart';
 
+import 'extension.dart';
 import 'mocks.dart';
 
 void main() {
+  Future<void> buildWidget(WidgetTester tester) async {
+    await tester.pumpWidget(
+      ProviderScope(
+        overrides: [
+          gitHubRepoServiceProvider.overrideWithValue(MockGitHubRepoService()),
+        ],
+        child: const MaterialApp(
+          localizationsDelegates: [S.delegate],
+          home: Scaffold(
+            body: RepoSearchBar(),
+          ),
+        ),
+      ),
+    );
+  }
+
   testWidgets('shows search bar correctly', (WidgetTester tester) async {
-    await _buildWidget(tester);
+    await buildWidget(tester);
 
     expect(find.byType(SearchBar), findsOneWidget);
     expect(find.byType(Icon), findsOneWidget);
@@ -18,7 +34,7 @@ void main() {
 
   testWidgets('enter text updates the query provider',
       (WidgetTester tester) async {
-    await _buildWidget(tester);
+    await buildWidget(tester);
 
     await tester.enterText(find.byType(TextField), 'flutter');
     await tester.testTextInput.receiveAction(TextInputAction.search);
@@ -31,7 +47,7 @@ void main() {
 
   testWidgets('clear button clears the query provider',
       (WidgetTester tester) async {
-    await _buildWidget(tester);
+    await buildWidget(tester);
 
     await tester.enterText(find.byType(TextField), 'flutter');
     await tester.testTextInput.receiveAction(TextInputAction.search);
@@ -46,27 +62,4 @@ void main() {
     expect(query, isNull);
     expect(find.text('flutter'), findsNothing);
   });
-}
-
-Future<void> _buildWidget(WidgetTester tester) async {
-  await tester.pumpWidget(
-    ProviderScope(
-      overrides: [
-        gitHubRepoServiceProvider.overrideWithValue(MockGitHubRepoService()),
-      ],
-      child: const MaterialApp(
-        localizationsDelegates: [S.delegate],
-        home: Scaffold(
-          body: RepoSearchBar(),
-        ),
-      ),
-    ),
-  );
-}
-
-extension _ReadQuery on WidgetTester {
-  String? readQuery() {
-    return ProviderScope.containerOf(element(find.byType(RepoSearchBar)))
-        .read(queryProvider);
-  }
 }
